+++ spec/controllers/discovery_controller_spec.rb
@@ -20,12 +20,10 @@ describe Doorkeeper::OpenidConnect::DiscoveryController, type: :controller do
         'scopes_supported' => ['openid'],
 
         'response_types_supported' => ['code'],
-        'response_modes_supported' => ['query', 'fragment'],
+        'response_modes_supported' => %w[query fragment],
+        'grant_types_supported' => %w[authorization_code client_credentials],
 
-        'token_endpoint_auth_methods_supported' => [
-          'client_secret_basic',
-          'client_secret_post',
-        ],
+        'token_endpoint_auth_methods_supported' => %w[client_secret_basic client_secret_post],
 
         'subject_types_supported' => [
           'public',
@@ -55,6 +53,17 @@ describe Doorkeeper::OpenidConnect::DiscoveryController, type: :controller do
           user_info_response
         ],
       }.sort)
+    end
+
+    context 'when refresh_token grant type is enabled' do
+      before { Doorkeeper.configure { use_refresh_token } }
+
+      it 'add refresh_token to grant_types_supported' do
+        get :provider
+        data = JSON.parse(response.body)
+
+        expect(data['grant_types_supported']).to eq %w[authorization_code client_credentials refresh_token]
+      end
     end
 
     it 'uses the protocol option for generating URLs' do