+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - [#114] Fix user_info endpoint when used in api mode
+- [#112] Add grant_types_supported to discovery response
 
 ## v1.7.2 (2020-05-20)